add filteroptions to BooleanField and refactor
